+++ R/utils.R
@@ -297,6 +297,14 @@ drop_elements <- function(x, n) {
   x[seq_along(x) > n]
 }
 
+#' Ask for Confirmation on Large Visit Levels
+#'
+#' @description Ask the user for confirmation if there are too many visit levels
+#' for non-spatial covariance structure in interactive sessions.
+#'
+#' @param x (`numeric`)\cr number of visit levels.
+#'
+#' @keywords internal
 h_confirm_large_levels <- function(x) {
   checkmate::assert_numeric(x)
   allowed_lvls <- x <= getOption("mmrm.max_visits", 100)
@@ -304,11 +312,17 @@ h_confirm_large_levels <- function(x) {
     return(TRUE)
   }
   if (!interactive()) {
-    stop("Visit Level larger than 100.")
+    stop("Visit levels too large!", call. = FALSE)
   }
-  proceed <- utils::askYesNo("Visit levels is larger than 100. This requires large memory. Are you sure to continue?")
-  if (!proceed) {
-    stop("Visit Level larger than 100.")
+  proceed <- utils::askYesNo(
+    paste(
+      "Visit levels is larger possibly too large.",
+      "This requires large memory. Are you sure to continue?",
+      collapse = " "
+    )
+  )
+  if (!identical(proceed, TRUE)) {
+    stop("Visit levels too large!", call. = FALSE)
   }
   return(TRUE)
 }